requirements.txt
fastapi
paddleocr
pillow
python-gtts
playsound
hbcvt
uvicorn